Unverify region violating clients only once
We check over our entire clients list to see if they can operate within
the exits current region lock. Right now we will check and unverify even
users who have already been denied. Resulting in a constant continued
rejection for that same user.

This is inefficient and is resolved in this patch.