Fix forwarding for phone clients
To access antennas on these ports we need to use the br-pbs bridge
the ports are attached to, not include the ports directly.

I'm unsure of the security implications of this, especially since .30
isn't a blocked address. Could this potentially allow for dashboard
access since we're not subject to our own firewall? This will need
to be tested before release.