Use AtomicBool for is_gateway rather than settings
We where skipping deserializing the is_gateway setting because we
didn't actually want persistance, we wanted an atomic in memory
thread safe data store.

An atomic bool performs this job with no issues.

We can also remove an actix message for the gateway client corner
case in billing by using this same atomic bool strategy. Allowing
us to move forward our goal of eventually removing actix actors.

Finally this fixes a billing corner case where a node that is
a direct exit peer would not identify itself as a gateway when it
comes to setting prices.