Remove withdraw_eth endpoint
The withdraw_eth endpoint was not actually being called by the dashboard
or anywhere as far as I'm aware. The supported way to withrdaw all your
eth is to use the withdraw_all endpoint, which cleans up the eth on it's
way out leaving both the xdai and eth addresses empty.