1/3 Async Token Bridge
This refactors the entire TokenBridge module to use async/await. This patch
is obviously useless until we get an async/await execution environment working
in the second patch in this series.

There is significant risk of semantics change with this refactor and testing
will be required. Likewise the operative flow of the program is changed enough
that some of things we're doing no longer make sense. There are errors we can
handle now and things we can keep track of (like the actual amount exchanged)