Move tunnel GC to fast loop
Tunnel GC is the only component in the Rita slow loop that requires
access to actix messages. By moving it into the fast loop we can
convert the entire slow loop rather than duplicating it to provide
both actix contexts.

This may cause issues with running the GC code once every 5 seconds
rather than one a minute. But I think it's fast enough not to worry
considering that most clients only have a couple of tunnels.