Refine token bridge logic
This patch reformats the token bridge logic to do away with hardcoded
reserve and minimum to exchange amounts. In there place there are dynamicly
computed amounts based on known operation costs and the current gas price.

Likewise the lower level logic of the token bridge is changed to send transactions
more robustly but never sending transactions that exceed the total amount of ETH in
our balance to pay for gas.

Finally the logic has been re-factored to make more sense in an async-await context
we now have access to variables that where previously too difficult to capture and
the execution structure can be significantly reduced in complexity now that we don't
have more than one future flying around at a time modifying the same state.