Tune Cake overhead and add to Client Exit Tunnel
This adds Cake on ingress to the client exit tunnel and tunes the overhead
params to hopefully provide a better experience at both levels. The additional
diffserv parameters should be helpful on the exit client tunnel.

I don't know if they will actually do anything on per-hop tunnels as it's
not totally clear to me if Cake is doing diffserv based on packet tags, it's
own mesurement or both.