Revert "Unmonitor tunnels when re-creating them"
After thinking on it. The core of this issue is in babel. While this
patch hide that babel issue it also creates a race condition where
we could unlisten the tunnel after attempting to listen onto the
'new' tunnel. Resulting in an abandoned tunnel. GC will clean it
up with recent changes, but why bother if we can fix it at the babel
level

This reverts commit 629b07f59657bf2af902cf947ebb1b89c5414d39.