Reuse database connection for region enforcement
The exit previously allocated another database connection for exit
region enforcmenet. This had no particular advantages as it was all
performed in sync code while the caller was already holding a connection.

This lead to the operation failing in environments with low available
connections and alerted me to the issue.

This patch does the obvious and has region enforcement reuse the connection
it's caller is holding anyways.