Correct tunnel GC behavior
This was quite the doozy of an incident. This change occured over
two parts. One I made a change that checked if babel tunnels where
up and removed the tunnel if they where not. Seperately I moved the
tunnel GC operation (aka when the previous change was called)
into the fast loop.

The kicker here is that it takes some time for tunnels to come up
after creation, exactly how long is based on the link quality and
router being used.

If this change had been left in the slow loop (60 seconds) it would
have never been relevant. But instead it was placed in the fast loop
(5 seconds) and then tested mostly with fast routers. That where able
to bring their tunnels up within 5 seconds consistently.

As a final complication if the other peer was not suffering form the
same issue no problem arises. So if either side is fast no issues.

This long combination of conditions hid the race condition presented
by the combined two changes above and allowed it to reach large scale
production.

This new patch introduces a new variable on the tunnel struct which
explicitly allows us to avoid this condition and provide a grace period
I've also taken the liberty of documenting the involved code in
excruciating detail to avoid this issue again.