Add ChangeOperatorAddress operator action
This adds a simple operator action for superuser
use by the operator tools. Pushing the router to a
new operator address or removing it from a network
entierly.

I believe this is safe to add as a command becuase it
will be access controlled on the ops tools side and it
does not modify the reporting address url which is still
ssl secured (therefore updates are signed)

The last change here is expanded parsing to make it easier to
send this command into the backend from the frontend