Add ChangeReleaseFeedAndUpdate Operator action
This operator action allows for the easy changing of the release feed
and updating of an individual router as a single operation. This better
fits our current model of staged releases where we update specific routers
in order to solve bugs or test new releases.

The old strategy of having all devices on a global feed is good for global
updates but not so great for updating a single device or updating a single
network. This should provide more flexibility by far.

Once again this will have to be access controlled on the operator tools side