Double gas price multiplier for bridge transactions
Due to recent gas price spikes on Ethereum the median gas price
for the last 100 blocks is far too low. Not because we will actually
pay that much more but becase everyone is advertising a willingness
to pay more.

The lowest price auction method of setting block wide gas price means
that miners are collecting many transactions of higher value and then
inserting lower value ones only at the end. A few lucky people get in
and bring the gas price down for everyone resulting in a block median
much lower than the actual bidding price.

Hopefully this works well with the gas price send limiter, which should
keep us from trying to pay more in gas than we actually have in funds.