Add configurable remote logging to Rita Exit
As time goes on we're getting more on site exits in more locations.
This brings the logging model closer to routers, bandwidth usage
matters and it's infeasible to whitelist them on the log collecting
server side.

This should provide the flexibility to use the right tool for the job
depending on where the server is and how much we think it's worth
it to maintain the syslog seup.