Fix withdraws for the unknown amount_exchanged state
If your uniswap transaction goes through within a few seconds of being
issued then the bridge will successfully capture the event and know
exactly how much was exchanged. This is presented in dai wei. But
if we don't capture that value because transactions are moving slowly
we fall back to an exchange rate computed estimation.

When I lasted edited this code I failed to convert from DAI to DAI wei
for the threshold comparison, resulting in an error where withdraws
would always halt at the last step.