Combine withdraw handlers
Having seperate handlers for withdraw and withdraw_all is just asking
for edge cases where a 'withdraw' is not quite all of the funds in the
router but still too expensive to process ignoring fees.

This refactor combines the backend of both endpoints so that fees are
always computed and handled if required.