Reduce tunnel opening to a single operation
In an effort to reduce the overall complexity of error handling in
Tunnel manager and generally the readability of the code new() and
open() have been combined into a single method that performs both
functions.

While this makes for an individually more complex function there is
now a single scope for handling errors and no more 'gotcha's if you
call new everything gets taken care of.