Retry tunnel opening on duplicate
This makes two changes to the tunnel opening interactions with the
ip command.

First we previously ignored tunnels that where already open as an
error. This is a hold over from when tunnel opening was not imdepotent
and is no longer required and in fact actively playing into a bug where
a tunnel that does not exist in 'ip a' output or actually work has been
opened.

What we do now is catch an 'exists' error and increment the tunnel
number in an attempt to get out of the way of the invisible tunnel.

The other change made here is that instead of crashing on a tunnel
open failure we will now try again. I don't think this was actually
happening as processes have not been observed to restart often.

Either way it's better to handle it live rather than by crashing