Remove Failure from Kernel interface and Types
This follows the trend of removing the now depricated failure crate
from as many places as is feasible. It is very much not feasible to
do so in the primary Rita crate at this time mostly due to how tightly
integrated failure is in the futures01 stack.

But at least we've got more specific errors where possible now that
it's become much easier to implement the std::error::Error trait