Refactor PaymentValidator full node contact protection
This change adds a probably overengineered system to meet two
requirements.

1) PaymentValidator should play nice with the new thread execution
environment and not panic directly, instead calling the main actix
loop before exiting

2) We add a long running record of all failed transactions and their
reasonings that we use to ensure that we do panic if we can't contact
full nodes for a long period but don't panic if a single transaction
is invalid and can't be checked.

to removing that protection given that our new thread based system
where we spawn a new actix executor on each loop may never produce it.