Seperate outgoing and incoming payment timeouts
This essentially reverts part 2 of fd3d8b4d2efb5231e3a501088c17bf77f07fd3b8

After seeing another related incident in production I had the realization that
the checked variable itself needed to be computed more carefully.

We now define a transaction as checked if we manage to reach a full node at all
when we got to check it. This dramatically reduces the chance that we hit the
payment failure panic condition in the first place and is much easier to express
than an entire payment failure introspection system.

The other change here is to introduce a much shorter payment send timeout which
is roughly an average amount of time for a tx to enter the xdai blockchain after
we send it. It's my hope that this won't have too many false positives and will
manage to save our bacon from enforcement on very active nodes.