Add Tunnel gc condition for no-hanshakes
After observing a tunnel mis-negotiation in production I am considering
a race condition where tunnels may end up mis-negotiated. While I've
done my best to design the tunnel opening process to make this impossible
I can't pin down what could cause this at the moment.

So we have an open tunnel, properly connected to babel, that's not expired
becuase we're still recieving messages from the other side. The only way
to determine that it's not working is to establish that Wireguard has failed
to handshake successfully.

Previously I had been trying to avoid shelling out to wireguard for
every tunnel in our tunnels list every time garbage collection was run
but if we need to check this value every time anyways there's no avoiding
it. With that in mind I have refactored the conditionals into a match
statement that is much easier to read.