Don't run TunnelGC more often than once a minute
Changes last release required TunnelGC to be moved into the fast loop
at least until TunnelManager goes through it's async/await refactor.

After a production incident the code here was patched to be comfortable
running much faster than before. But some additional complexity was
brought in to optimize.

In our last patch we conceded the need to shell out for every tunnel
during the GC operation, this is not optimal to do at all much less once
for every tunnel every 5 seconds. This patch reduces the run speed of
TunnelGC so that it doesn't run every time it's called and should do
the trick for now.