Allow registration to no-auth exits without auth details
Previously the user was required to fill in authentication details
in order to register to exits that don't require authentication.

This is obviously an oversight, especially since exits no longer
manage low balance notifications.

Previously this had been hidden by a front end check that was removed
during a refactor.