Move TokenContractAddresses struct to auto-bridge
This patch adds xdai_home_helper_address to the token bridge addresses
struct. In the future this address will be used to faciliate working
withdraws right now it's just the inspiration for a bit of a refactor.

Having the addresses struct defined in settings means we have to feed in
the token contract addresses as easily confused and mixed up unnamed
arguments, this organization is less likley to result in a fatal mixup
while changing these constants.

This also allows us to consolidate all definition of the token bridge
addresses as constants into one location in the repo.