Support for xDAI bridge withdraws
Literally less than 24 hours after my last set of bridge improvements
the xDAI team broke withdraws by updating their contract to have the
user pay for the gas. This makes sense as a feature but I remain pissed
they did it with less than 24 hours of warning.

This patch collects the signatures using a helper contract on xDAI then
submits them to the Ethereum contrcat in order to unlock the funds.

I suspect this may require a re-consideration of the reserve amount as
well as various other tuning. But here is basic support to expand upon.