Default route setting refactor
This patch modifies the default route handling significantly.

Previously the default route was stored simply as an array of strings
this actually worked remarkably well for what it was. But in some cases
it would replay tokens on the ip route that would intentionally make the
link not work (for example 'linkdown') preventing nodes from coming back
up.

This new version provides a much more in depth parser and explicitly
handles previously implicit edge-cases. For example when no default
route was set previously we just got an empty array and the command
would fail to run, this would implicitly disable default route setting
on non-gateways. Now we can simply detect this by storing an
Option<DefaultRoute> struct.

I do have some concerns about this code, namely it's immaturity and the
possibility of unexpected ip route outputs causing failures on gateways
that currently work. To that end I've beefed up the tests as much as
possible and also tried to design the parser to be very resistant to
strange tokens.

The last and only actual routing taable difference between the previous
design and this one is that we now add 'proto static' to the routes we
set this tag ensures that scripts that respect conventions don't mess
with our routes.