Heartbeat info caching
We've been having some issues with Heartbeat reliability in production,
I suspect it's a combination of DNS, routes, babel parser failures and
actual crashes. But this patch will at least clear up the lack of
heartbeat sending due to spurious parsing failures.

Instead of only caching DNS results we now cache all info required to
send a heartbeat, if we fail to get any of it we will still retrieve the
required info from the cache and send it along.