Attach a full identity to Wyre request messages
Previously I did not properly consider what it would take to authenicate
individual users for higher trust payment onboarding flows. While I can
move forward with only a deposit address it will require an inefficient
lookup on the Operator tools side of things. One that I would prefer not
to have to deal with in the future. Hence this change.