Move operator fee to an accumulation based system
Really dramatic operator tool fee changes, like from $0/mo to $100 a
month would previously try and backpay the entire new fee amount from
the time the router was last booted. This is obviously quite a lot of
money and not intended behavior.

The operator fee modules entire system of determining how much to pay
purely based on the last payment time clock was 'clever' in that it
minimized fee amounts lost to sub-second truncation caused by elapsed()
but this flaw and others like it are simply too much to deal with.

In the very worst case this change will lose 12 hours worth of fees a
month, given that every time elapsed() is called it's the smallest
possible amount below the next second. That's a little over 1.5% which
I think is acceptible for the stability.