Delay the deletion of wg interfaces
So this is a race condition come back to bite us.

More than a year ago this code was modified to deal with a change in how
the upstream kernel handled the default setting of rp_filter. Once this
behavior made it into openwrt we started seeing a lot of random tunnels
not working.

After locating a sudomesh issue I determined this was because we where
deleting the underlying interface before removing the interface from
babel. Which results in netlink issues, we moved to unmonitoring in
bable before we deleted the interface and the issue seemed solved.

A little over 9 months ago bug reports where recieved about nodes with
high uptime no longer accepting new peers. I observed in this case some
tunnels where not 'up' in babel. This patch fixes that issue (hopefully)
as a race condition in our fix. Where Babel may not have finished
unmonitoring the interface by the time we go off and delete it
ourselves.

Hopefully combined with some fixes to babel itself imported form
upstream we'll have some luck putting this issue to bed.