Exits should not send low balance notifications
I believe we had an incident this morning centered around this issue
although I'm still not totally sure I have it pinned down. Anyways
duplicating this functionality is only prone to error.

Note that this does not remove the required fields from the exit
database schema. I'd prefer to keep that static as we refactor the
exits.