Upgrade rita deps and resolve clippy nits
This resolves several new clippy nits and upgrades us to the latest
actix-web. I'm not really sure if using the beta actix is the right
choice for a high stability application like Rita, but we're somewhat
pressed for options due to other dependency tree issues upstream with
web30 and prost. An issue which will make it's way down into this
codebase eventaully anyways.

Going to see how this bakes, the other upgrades are pretty simple const
generics improvements.