Reduce operator checking frequency
This is a reduction of the checkin frequency from every 5 seconds to
once every minute, which logic to increase that value arbitrarily. This
is more efficient and will use less bandwdith on the part of the user as
there's really no reason this info needs to be updated so often.