Splitting rita into seperate crates in order to increase modularity and open source a portion
Rita was split into 4 different crates each of which call the settings crate. The settings
crate was modified to include lazy statics that allow for global mutation of a reference dynamic-
ally. This was done in order to properly split up the crates while somewhat preserving modularity.
Additionally, rita_common was modified to include some references such as ki in order to allow
individual crates to import the neccessary information.

All binaries were modified to remove the unused imports and local lazy
static variables that are now hosted in settings.

This refactor introduced a lot of bugs that were spotted by the integration and unit tests. The
full test suite passes, but there definitely may still be bugs/undesired behavior occurring.

Finally, the read me was modified to reflect a possible oversight that running integration test
will not actually compare to local files and the boilerplate rita tower
code was integrated into this patch.