Implementing rita uptime as a lazy static struct
Due to how lazy static is unsafe, we need a read write lock in order
to allow for safe handling between threads. Thus, there is a global
reference to a struct containing the duration type that is serialized
for the frontend dashboard. OperatorCheckin struct was modified to
include the duration type field. A mod test was written as well just
to verify that the struct does indeed work as intented. There are no
known bugs or issues at this time.