Move antenna forwader start into client.rs
Since the antenna forwarder is so independent from the rest of Rita
there's no reason to start it in the Rtia common loop, much less have a
tracker that keeps checking if we have or have not started it.

This commit moves that startup sequence to the client binary which will
also allow for customizations there in the future.