Update the Gitlab CI definition
This patch gets the CI tests working again, but not the integration
tests. So we have some nice compile and lint checks but getting
Wireguard working in Gitlab ci is proving more challenging than it was
in Travis or github Actions. At this stage I believe I'm correctly
building the kernel module itself but for some reason can't load it.