No sensors is a valid outcome
CI test machines have no sensors, this is a valid response that should
not fail the tests. This entire issue really calls into question the
wisdom of testing these hardware info functions on just the local
machine and expecting correct outcomes. They won't succeed on Windows if
a developer was ever developing there and there's a huge breadth of
responses even in the Linux or Unix like world.