Update x86 build scripts
This patch updates linux_build_static which is used for building generic
linux versions of rita and rita_exit, the new strategy builds a docker
container with the required dependencies on top of the cross gnu
container.

This binary will work fine on most Linux distributions but not OpenWRT,
for that case a new openwrt_build script is provided specifically for
x86_64 machines, this only builds Rita not Rita Exit.

This eliminates the use of the (very old) Rust musl builder and
standardizes us on the use of either Cross or OpenWRT toolchains across
all Althea software.