Reduce logging lines to 1k
For Beta 17 we attempted to update to 10,000 lines of logging per
compressed log upload. This works fine for simple test cases lacking
real data, but when used with real data the payload becomes too large at
some point. Determining if this issue is on the client or server side
has been difficult so for the time being we'll just roll back the change