Removed get_port recursion and PortCallBack. Implemented new port management in Tunnel Manager.
Now all ports are in a vector, we provide a port when requested and move this port
to the end of the vector. When we come back around to this port, it is either successfully
allocated or is dropped, in which case it can be reused.

This fixes the previous PortCallBack issue in sync hello management in the previous commit