Fix test execution in auto-bridge
This patch fixes actix executor problems with the tests in this module.
Since these tests don't run by default and in this case the API change
was breaking but did not generate compile errors this went un-noticed
for some time.

A good next step here would be to add tests that check uniswap balances
or some other basic async test, which would flag issues like this in the
future.