Improving the reliability of withdrawal process by adding looping code to check events on the xdai blockchain.
In this commit, we look at a user entered block range, parse the data field to check and see
if there are any withdrawal events related to us. This allows us to find withdrawals to
assist debugging, or to find withdrawals to push along. This allows us to pay fees when
needed to unlock funds or to rescue stuck funds.

Previously, our withdrawal process was unreliable, this commit is one step towards
adding to the reliablity of the withdrawal process.