Don't use settings getters and setters
Due to the moving semantics of the borrow cheker using this pattern
involves moving data, then moving it back in using the setter before we
can save it.

Considering our usecase this caused a lot of extra lines and more than a
little confusion about what data we where editing exactly.

Previously these getters and setters had the advatange of providing
scoped locking, but that's no longer the case in the newly simplified
settings module.