Changing default wg private key path
The dfeault wg_private_key path has been a blank string for a pretty
long time. This is because all routers have this data set in the default
config and the value was added simply to satisfy the 'default' trait
requirements.

Regardless of the fact that it works it's a bad default value, as the
code can never actually run using this value. This patch correctst that
issue.