Modification to create wireguard for better debugging
Previously, the create wire guard key function would unwrap the result
of key generation. However, this made it rather unobvious as to why
the code was breaking when the code had all correct information. Although,
there could be an error with incorrect information at least now the program-
mer or installer will hopefully check to see if wireguard is now installed!