Fixed bugs in previous 'Remove Uniswap' commit
Added unit tests to check the withdrawal process at all steps. This includes the setup
of withdrawal information in the lazy static variable, transfer of funds with the relayTokens contract call
and the unlocking of funds by first simulating the transaction (with the relayedMessage contract call) and then
confirming the unlock with the executeSignatures contract call.

New deposit flow and detailedState rendering on the frontend still need to be tested.