Unmonitor tunnels syncronously
This removes the async tunnel monitoring and all futures based babel
from tunnel monitor. This is an aggressive change, but I believe it will
help reduce instances of a known bug around not deleting interfaces
before de-allocating them. Babel has a cache of interface details that
can become corrupted if the underlying interface is deleted before babel
is detached. This change makes it much easier to say with confidence
that that has occured. It does come at the cost of potentially strange
behavior of connections to babel are blocking. But we'll take that trade
at this point.