Implemented better exit migration and failover
When babel decides when to switch between exits, this may be less productive as everytime we
switch, wireguard takes ~1min to renegotiate. This is detrimentral when various exits are 'better'
for short periods, causing downtime when in fact we could have stayed with the same exit.

By moving this logic into rita, we are able to make informed decisions about when to switch exits.
We have a 'timer' of fifteen minutes; the minimum time we need to wait before considering to switch. After
15 mins, we decide whether to switch or not based on babel metrics, how stable they are and how much better
they are than our current exit. This deliberated decision to switch helps prevent route flapping and ensure stable
exit connections.