Remove bogus features
When we split the single rita crate into three components a lot of the
features where copied over, and from there features where added to teh
list in the rita_bin Cargo.toml based on the other tomls, not based on
if those features where ever actually used. This patch resolves the
duplication.