Fixing Peer Listener bug unable to contact manual peers
Perviously peer listener looked for peer information and udp sockets of node
that sent out im here messages. However, when trying to contact manual peers, we would not
have this information in a hashmap and thus panic. Now, manaual peers reverted to the previous HHTP
hello using the hello handler, fixing this issue.