Moved Rita's update system from ash script to rust kernel interface
Instead of using opkg in the script, we now primarily use sysugrades to perform upgrades,
This helps us with space reduction on the router. Operator tools has the oppurtunity to use
both sysupgrade or opkg by specifying the download url of the firmware image or the url to
add to customfeeds.conf for opkg, along with a vector of opkg instructions

On the router dashboard side, updating is simplified. Now the user clicks an update button
and operator tools tells the router which version of rita to sysupgrade to, or which packages
to install via opkg