Resolve token bridge deadlock
The detailed token bridge state lock was being opened for writing twice
in the same context. First to set the withdraw state and second to
update the detailed state at the end of the xdai_bridge function.

This caused the function to deadlock and stop the slow loop from
continuing to run.

This patch resolves the by scoping the first write lock and also
performs a minor cleanup of the get_core() function which was only being
used in contexts where another read from the settings lock was not
needed. This change is purely cosmetic, there was no potential to
deadlock there.