Finish Rita Client traffic watcher async/await migration
During the exit manager upgrade this file was mostly migrated. But a
single actix endpoint was left to get the exit destination price. Since
the actix endpoint was simply reading from the lock anyways finishing
the migration was trivial and had the knock on effect of cleaning up the
code around each of it's usages.