More protections in make_payments endpoint
payment_validator correctly discards payments that are not actually
being sent to us, but the make_payments endpoint will still return 200
to the confused node, causing it to believe the payment was successful.

This new error code won't do anything to actually resolve the issue, but
will at least provide an error for that other node to follow up on. We
also log this error so it is visible to the user.