Added a blacklist for rogue exit ip addrs
It is possible that we have a rogue exit ip that collides in our connected subnet,
leading to a blackhole attack, as exit switcher will continuously try to connect to this. This blacklist will
blacklist an ip if we are not able to decrypt the ExitStatus of an exit in our ip.

Also it may be possible that we get no reponse to our exit status query, meaning the exit is slow or there is
a router in our subnet that we think is a exit. In this case, we have a penalty system that tries to connect
to this exit for 'n' ticks, after which this ip is added to the blacklist and is deregarded