Implement a settings adaptor
In order to let a wrapping binary work with the existing settings usage within rita_common, but to add extra fields, we can use an adaptor to redirect disk and cache io calls to the wrapper.

This commit adds a WrappedSettingsAdaptor that allows settings to be handled by a higher layer that wraps RitaClientSettings and adds its own additional fields. The higher layer is responsible for the actual read/write of settings to disk.

This also cleaned up a bunch of strange settings read/writes in the dashboard http handlers. In testing it seems to be more snappy.

This has been tested with a wrapping binary for verification.