This commit does some analysis to find out what a good CLOSE_THRES multiplier would be post EIP1559 to prevent false enforcements, when in the worst case a router runs 10x slower than another router that it recieves pay from
In the average case where gas prices remain relatively stable, averaging gas values over a period of time lead to
the two routers agreeing on gas prices more than just using the latest gas value. But in the worst case this averaging can
lead to a 20x disagreement on gas pricing, compared to just using latest, which has a worst case disagreement of 4x.
With this, CLOSE_THRES multiplier is set to a conservative value of 10, using latest gas price instead of averages to pay and close thres
calculation

Other cleanup was done in this commit such as moving transient values from settings struct to a lazy static in blockchain oracle
and removing other redundant code for the post EIP1559 transistion