Fix and clean up usage tracker
This patch is a general cleanup of the usage tracker. The usage tracker
had left over cruft from two different migrations, one to async/await
and the on disk migrations from flatfile storage to compressed storage
plus the rename of the misspelled internal struct.

The combined cruft of these migrations resulted in a very simple error.
previously handle payments was called periodicly by Actix using functions,
as a two part migration payments would be queued onto the payments queue
when sent from an async/await using payments function and then dequeued
when an old futures payment function was called.

Since this file was totally updated to async/await the intermediary call
and the queue where no longer needed but where still being used.
Previnting payment data from ever reaching the disk or the user.

Another issue that became apparent is that json is a rather memory
hungry format and on my test glb1300 the combined payment history for
the clatskanie gateway would cause Rita to be killed by the OOM just
loading the data.

This patch includes a partial fix that takes advantage of streaming
reads and writes to load the data. But that is not enough to allow
returning the data as json from the request endpoint. In order to
mitigate this problem I'm reducing the maximum entries significnatly
for transactions.