Reduce default xdai fee multiplier to 3000
Payments on xDai are now far too large since they are enforced by the
min fee value post EIP-1559.

This reduction to the payment multiplier will increase the payment
frequency by increasing the fraction of a payment that is allowed to be
fees from 1/6000 to 1/3000.

This is safe as it is a 2x reduction that is within the 4x CLOSE_THRESHOLD_MULT
currently deployed on Beta18RC9 and earlier builds. This means routers
with this value and routers with the old value of XDAI_FEE_MULTIPLIER
can pay each other without enforcing, the margin of error is increased
for enforcmenet from an older version on this newer version since it
will pay more often and reduced in the reverse direction.

Since we just incresed the CLOSE_THRESHOLD_MULT in the last
commit to 10 this more than restores that lost margin preventing this
new version for enforcing on the old.