Provided a safe shutdown method for rita client and exit
Added a handler for SIGTERM, SIGINT, and SIGHUP to save debts, usage and rita settings before shutdown
Previously, during a restart, we would lose all this information since it wasnt being saved to disk. On the exit
for example, many clients can have their debt and credit balances wiped out from the debt keeper from a restart.
This can amount to hundreds of dollars worth of errors.
This commit adds a handler to fix these issues, both for exit and client