Set a minimum gas price to 1Gwei
We don't want gas prices to go below 1gwei or we will see payments
occuring too frequently, the multipler controls how often we actually
pay and prevents our overpayment from actually impacting us in terms of
price. We could just have a higher multiplier, but it's better for gas
prices to be generally stable to keep payments the same size. With this
design payment size will go up if it's really needed but we have a lot
of cushion before that's required.