Update web30
This new version of web30 fundamentally changes the behavior of the
eth_gas_price call. Instead of returning the maximum of eth_gas_price
and the block min fee it now returns only the block min fee when
available. This is a decision driven by the observation that on Ethereum
and other fee constrained chains the actual fee is effectively the min
fee and on chains like xDai that are running EIP1559 but not fee
constrained the min fee is much lower than the average fee. Therefore
it's optimal to always select the min fee as our value.