Re-add min_gas setting
Post EIP1559 on xdai (gnosis chain) I believed the min gas setting could be safely removed,
after all with the implementation of EIP1559 it was no longer needed to
have the validators conspire and set a minimum gas fee there was now a
defined value to use.

This ended up being wrong it seems, as the EIP1559 block base fee is
based on how full the block is, and since xdai blocks are essentially
always empty the value is extremely low, well below one gwei. Instead of
modifying the EIP1559 consensus parameters the xdai validators seem to
have instead just re-implemented the same min fee logic as before but
now with a higher value.

What this patch does is restore the min_fee setting and provides an
operator action for updating it as I suspect we will be doing this
often.